# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhatPay API",
    "description": "Plataforma de pagos peer-to-peer integrada con WhatsApp para Chile.\n\n## Autenticación\nLa mayoría de endpoints requieren un JWT Bearer token obtenido en `POST /api/v1/users/login`.\nLos endpoints admin requieren el header `x-admin-key`.\n\n## Flujo típico\n1. Registro vía bot WhatsApp (conversacional) o `POST /api/v1/users/register`\n2. `POST /api/v1/users/login` → PIN → JWT (30 min)\n3. `POST /api/v1/topup/webpay` → recarga wallet con Transbank\n4. `POST /api/v1/payments/pay` → envía pago P2P\n5. `POST /api/v1/payments/links` → crea link de cobro para comercios\n\n## Niveles KYC\n| Nivel | Límite por tx | Límite mensual |\n|-------|--------------|----------------|\n| `BASIC` | $50.000 | $200.000 |\n| `INTERMEDIATE` | $500.000 | $2.000.000 |\n| `FULL` | $2.000.000 | Sin límite |\n\n## Seguridad\n- PIN almacenado como bcrypt hash (cost 12)\n- RUT almacenado como HMAC-SHA256 (nunca en claro)\n- Bloqueo tras 3 intentos fallidos de PIN (15 min)\n- Rate limiting por IP en todos los endpoints\n- Todas las transferencias usan SELECT FOR UPDATE (sin double-spending)\n",
    "version": "0.1.0",
    "contact": {
      "name": "Pablo Cussen",
      "url": "https://cussen.cl",
      "email": "pablo@cussen.cl"
    },
    "license": {
      "name": "Portfolio Project",
      "url": "https://cussen.cl/whatpay"
    }
  },
  "servers": [
    {
...
      "description": "Producción (GCP Cloud Run — Santiago)"
...
    {
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido en POST /api/v1/users/login"
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "API key de administrador"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Datos inválidos."
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/qr/generate": {
      "post": {
        "tags": [
          "qr"
        ],
        "summary": "Generar código QR",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "static",
                      "dynamic"
                    ]
                  },
                  "merchantId": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "integer"
                  },
                  "description": {
                    "type": "string"
                  },
                  "expiresInMinutes": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "QR generado",
            "content": {
              "application/json": {
                "example": {
                  "qr": {
                    "reference": "AB12CD34"
                  },
                  "qrPayload": "...",
                  "scanUrl": "/api/v1/qr/scan/AB12CD34"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/qr/my": {
      "get": {
        "tags": [
          "qr"
        ],
        "summary": "Mis códigos QR",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "QR codes",
            "content": {
              "application/json": {
                "example": {
                  "qrCodes": [],
                  "count": 0
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/qr/scan/{reference}": {
      "get": {
        "tags": [
          "qr"
        ],
        "summary": "Escanear QR (público)",
        "security": [],
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Referencia del QR (8 chars)"
          }
        ],
        "responses": {
          "200": {
            "description": "QR info",
            "content": {
              "application/json": {
                "example": {
                  "qr": {}
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/qr/{id}/use": {
      "post": {
        "tags": [
          "qr"
        ],
        "summary": "Marcar QR como usado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del QR"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transactionRef"
                ],
                "properties": {
                  "transactionRef": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usado",
            "content": {
              "application/json": {
                "example": {
                  "qr": {}
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/qr/{id}": {
      "delete": {
        "tags": [
          "qr"
        ],
        "summary": "Cancelar QR",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelado",
            "content": {
              "application/json": {
                "example": {
                  "message": "ok"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
...
    "/api/v1/scheduled-transfers": {
      "post": {
        "tags": [
          "scheduled-transfers"
        ],
        "summary": "Programar transferencia",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "receiverPhone",
                  "receiverName",
                  "amount",
                  "description",
                  "frequency",
                  "scheduledDate"
                ],
                "properties": {
                  "receiverPhone": {
                    "type": "string"
                  },
                  "receiverName": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "integer",
                    "minimum": 100
                  },
                  "description": {
                    "type": "string"
                  },
                  "frequency": {
                    "type": "string",
                    "enum": [
                      "once",
                      "weekly",
                      "biweekly",
                      "monthly"
                    ]
                  },
                  "scheduledDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "scheduledTime": {
                    "type": "string",
                    "example": "09:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Programada",
            "content": {
              "application/json": {
                "example": {
                  "transfer": {
                    "id": "stx_xxx",
                    "frequency": "monthly"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      },
      "get": {
        "tags": [
          "scheduled-transfers"
        ],
        "summary": "Mis transferencias programadas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transfers",
            "content": {
              "application/json": {
                "example": {
                  "transfers": [],
                  "count": 0
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/scheduled-transfers/{id}": {
      "get": {
        "tags": [
          "scheduled-transfers"
        ],
        "summary": "Detalle de transferencia",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Transfer",
            "content": {
              "application/json": {
                "example": {
                  "transfer": {}
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scheduled-transfers"
        ],
        "summary": "Cancelar transferencia",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelada",
            "content": {
              "application/json": {
                "example": {
                  "message": "ok"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
...
  }
}